Added settings panel to set the Data directory.
This settings panel is displayed on the File section of the main Blender
user preferences window and currently has the single option to set the
Jack Claw data directory. The Jack Claw data directory will be used to
import files that are not always placed in the same directory as the s3d
files, such as some reflection images used by the models.

A current issue that needs to be resolved is that the directory path is
not saved when Blender is closed

This settings panel will be a good place to have other options available
when needed in the future.