Image loading from Data/Texture when importing.
When loading images if the image file is not found in the same directory
as the s3d file, the add-on attempts to load it from the Data/Texture or
Data/Texture/Cubemap directory. These locations are found using the
user defined Jack Claw Data Directory setting that can be set using the
panel introduced in the previous commit.

Textures found in Data/Textures seem to mostly be files used for
multiple models such as reflection images.