The Jack Claw Data path can now be saved.
This is done using the python pickle module and loading/saving to a file
called 'settings.pkl' inside the same directory as the script files.

If there is a supported way to have persistent add-on settings within Blender
that I've not found yet, or is introduced in future Blender versions it would
probably be better to switch to that rather than using this self made method.